refactor: split V8Serializer into its own file
